Update Method TotalStockItem
Masih belum fix karena masih ada code yang belum selesai jika nanti sudah digunakan secara proper. Pada bagian item create/update/delete dan juga totalstock mengikuti dan menyeimbangkan hal tersebut.

+++ src/functions/items.js
@@ -1,7 +1,8 @@
 /** @format */
 
+import { CreateTotalStockItem, DeleteTotalStockByID, UpdateTotalStockByID } from '../functions/totalStockItem.js';
 import { FormatItemModel, ItemsModel, mongoose } from '../models/items.js';
-import { CheckingIsNilValue, CheckingKeyReq, CheckingKeyReqSyntax, CheckingObjectValue } from "../utils/utils.js";
+import { CheckingIsNilValue, CheckingKeyReq, CheckingKeyReqSyntax, CheckingObjectValue } from '../utils/utils.js';
 
 export const CreateItem = async (req, res) => {
   try {
@@ -28,10 +29,27 @@ export const CreateItem = async (req, res) => {
       itemSellPrice,
     });
 
-    return await newItem
+    let [isSuccessSaved, msgErr] = [false, 'err'];
+    await newItem
       .save()
-      .then((result) => res.status(201).json({ status: 'success', messages: `Berhasil menyimpan produk.` }))
-      .catch((err) => res.status(500).json({ status: 'failed', messages: `Gagal menyimpan produk. Catch: ${err}` }));
+      .then((result) => {
+        isSuccessSaved = true;
+      })
+      .catch((err) => {
+        [isSuccessSaved, msgErr] = [false, err];
+      });
+
+    if (!isSuccessSaved) {
+      return res.status(500).json({ status: 'failed', messages: `Catch: ${msgErr}` });
+    }
+
+    return await CreateTotalStockItem(req, res).then((result) => {
+      const [isSuccess, msg] = result;
+      if (!isSuccess) {
+        res.status(500).json({ status: 'failed', messages: `Gagal menyimpan total stock dan produk Catch: ${msg}` });
+      }
+      res.status(201).json({ status: 'success', messages: `Berhasil menyimpan total stock dan produk. ${msg}` });
+    });
   } catch (err) {
     return res.status(500).json({ status: 'failed', messages: `Gagal menyimpan produk. Function Catch: ${err}` });
   }
@@ -143,7 +161,7 @@ export const GetItemByID = async (req, res) => {
 
     documentsInDB = !documentsInDB ? await ItemsModel.findById(id) : documentsInDB;
 
-    if (documentsInDB.length < 1) {
+    if (!documentsInDB || documentsInDB.length < 1) {
       return res.status(404).json({ status: 'success', messages: `Tidak ada data.` });
     }
 
@@ -170,13 +188,13 @@ export const UpdateItemByID = async (req, res) => {
       return res.status(404).json({ status: 'failed', messages: `Format tidak sesuai!`, format: FormatItemModel });
     }
 
-    const isItemNameUsed = await ItemsModel.aggregate([{ $match: { itemName: itemName.toLowerCase() } }]);
+    // const isItemNameUsed = await ItemsModel.aggregate([{ $match: { itemName: itemName.toLowerCase() } }]);
 
-    if (isItemNameUsed.length >= 1) {
-      return res
-        .status(403)
-        .json({ status: 'failed', messages: `Nama produk sudah terdaftar! Silahkan untuk mengganti nama atau kategori.` });
-    }
+    // if (isItemNameUsed.length >= 1) {
+    //   return res
+    //     .status(403)
+    //     .json({ status: 'failed', messages: `Nama produk sudah terdaftar! Silahkan untuk mengganti nama atau kategori.` });
+    // }
 
     updateItem = CheckingObjectValue(updateItem, { itemName });
     updateItem = CheckingObjectValue(updateItem, { itemPublisher });
@@ -188,9 +206,18 @@ export const UpdateItemByID = async (req, res) => {
     updateItem = CheckingObjectValue(updateItem, { itemModalPrice });
     updateItem = CheckingObjectValue(updateItem, { itemSellPrice });
 
+    let [isSuccessSaved, msgErr] = [false, 'err'];
+    await UpdateTotalStockByID(req, res, documentsInDB, updateItem).then((result) => {
+      [isSuccessSaved, msgErr] = result;
+    });
+
+    if (!isSuccessSaved) {
+      return res.status(500).json({ status: 'failed', messages: `Catch: ${msgErr}` });
+    }
+
     return await ItemsModel.findByIdAndUpdate(id, updateItem)
-      .then((result) => res.status(200).json({ status: 'success', messages: `Berhasil memperbaharui data.` }))
-      .catch((err) => res.status(500).json({ status: 'failed', messages: `Gagal memperbaharui data. Function Catch: ${err}` }));
+      .then((result) => res.status(201).json({ status: 'success', messages: `Berhasil memperbaharui produk. ${msgErr}` }))
+      .catch((err) => res.status(500).json({ status: 'failed', messages: `Gagal memperbaharui produk Catch: ${msgErr}` }));
   } catch (err) {
     return res.status(500).json({ status: 'failed', messages: `Gagal mengambil produk. Function Catch: ${err}` });
   }
@@ -203,6 +230,12 @@ export const DeleteItemByID = async (req, res) => {
 
     if (!documentsInDB) {
       return res.status(404).json({ status: 'success', messages: `Tidak ada data.` });
+    }
+
+    let [isSuccessSaved, msgErr] = await DeleteTotalStockByID(req, res, true).then((result) => result);
+
+    if (!isSuccessSaved) {
+      return res.status(500).json({ status: 'failed', messages: `Catch: ${msgErr}` });
     }
 
     return await ItemsModel.findByIdAndRemove(id)
